fix: shortcut manager and menu crash + warning removal from Profile and Texture plotters

+++ Interface/ImguiMenu.cpp
@@ -241,7 +241,7 @@ static void ShowMenuFile() {
         ImGui::EndDisabled();
     }
     else if(ImGui::BeginMenu(ICON_FA_ARROW_CIRCLE_LEFT "  Load recent")){
-        for (size_t i = mApp->recentsList.size() - 1; i >= 0; i--) {
+        for (int i = static_cast<int>(mApp->recentsList.size()) - 1; i >= 0; i--) {
             if (ImGui::MenuItem(AbbreviateString(mApp->recentsList[i], 128))) {
                 std::string selection = mApp->recentsList[i];
                 auto common = [selection]() { DoLoadSelected(selection); };
